apps/batch/requirements.txt
openpyxl==3.1.5
pandas==2.2.3
psycopg2-binary==2.9.9
schedule==1.2.2
python-dotenv==1.0.1
requests==2.32.3
